Move registration info to helper
